[TrackingTools] Move public headers on to interface directory
